refactor: exception 패키지를 최상단에 위치
