+++ src/main/java/woowacourse/exception/shoppingcart/InvalidCartItemException.java
@@ -1,4 +1,4 @@
-package woowacourse.shoppingcart.exception;
+package woowacourse.exception.shoppingcart;
 
 public class InvalidCartItemException extends RuntimeException {
     public InvalidCartItemException() {
