+++ src/main/java/woowacourse/exception/shoppingcart/InvalidProductException.java
@@ -1,4 +1,4 @@
-package woowacourse.shoppingcart.exception;
+package woowacourse.exception.shoppingcart;
 
 public class InvalidProductException extends RuntimeException {
     public InvalidProductException() {
