+++ src/main/java/woowacourse/shoppingcart/dao/CustomerDao.java
@@ -13,7 +13,7 @@ import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
 import woowacourse.shoppingcart.domain.Customer;
-import woowacourse.shoppingcart.exception.InvalidCustomerException;
+import woowacourse.exception.shoppingcart.InvalidCustomerException;
 
 @Repository
 public class CustomerDao {
