+++ src/main/java/woowacourse/shoppingcart/ui/ControllerAdvice.java
@@ -11,11 +11,11 @@ import org.springframework.validation.FieldError;
 import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
-import woowacourse.shoppingcart.exception.InvalidCartItemException;
-import woowacourse.shoppingcart.exception.InvalidCustomerException;
-import woowacourse.shoppingcart.exception.InvalidOrderException;
-import woowacourse.shoppingcart.exception.InvalidProductException;
-import woowacourse.shoppingcart.exception.NotInCustomerCartItemException;
+import woowacourse.exception.shoppingcart.InvalidCartItemException;
+import woowacourse.exception.shoppingcart.InvalidCustomerException;
+import woowacourse.exception.shoppingcart.InvalidOrderException;
+import woowacourse.exception.shoppingcart.InvalidProductException;
+import woowacourse.exception.shoppingcart.NotInCustomerCartItemException;
 
 @RestControllerAdvice
 public class ControllerAdvice {
